refactor: move template.ts into shared #159
