+++ frontend/src/components/ProductDialog.vue
@@ -18,7 +18,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import { inject, Ref, ref } from "vue";
 import { saveProduct } from "../requests/product.ts";
 import { language } from "../../../shared/src/lang/lang.ts";
-import { interpolate } from "../lang/template.ts";
+import { interpolate } from "../../../shared/src/lang/template.ts";
 import { useProductStore } from "../store/productStore.ts";
 import { computed } from "@vue/reactivity";
 import { NewProduct, OptionalId, Product } from "../../../shared/src/types.ts";
