+++ frontend/src/components/products/ProductCategoryConfig.vue
@@ -21,7 +21,7 @@ import {
   ProductCategoryWithProducts,
 } from "../../../../shared/src/types";
 import { language } from "../../../../shared/src/lang/lang";
-import { interpolate } from "../../lang/template";
+import { interpolate } from "../../../../shared/src/lang/template";
 import ProductCategoryDialog from "../ProductCategoryDialog.vue";
 import ProductsTable from "./ProductsTable.vue";
 import { deleteProductCategory } from "../../requests/productCategory";
