+++ frontend/src/views/ShopView.vue
@@ -17,7 +17,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 <script setup lang="ts">
 import { onMounted, provide, ref, watch } from "vue";
 import { language } from "../../../shared/src/lang/lang.ts";
-import { interpolate } from "../lang/template";
+import { interpolate } from "../../../shared/src/lang/template.ts";
 import ShopItem from "../components/ShopItem.vue";
 import { useProductStore } from "../store/productStore.ts";
 import { useUserStore } from "../store/userStore.ts";
